+++ src/main/java/gift/controller/MemberController.java
@@ -10,9 +10,11 @@ import jakarta.validation.Valid;
 import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -42,11 +44,19 @@ public class MemberController {
 
     //TODO: 로그인 기능 -> 토큰을 반환
     @PostMapping("/login")
-    public ResponseEntity<JwtResponseDto> login(
+    public ResponseEntity<Object> login(
             @RequestBody @Valid MemberRequestDto memberRequestDto,
             HttpServletResponse response
     ){
-        String token = response.getHeader("Authorization");
+        //서버에 저장된 id-pw 쌍과 일치하는지 확인
+        if(!memberService.checkMember(memberRequestDto)){
+            //잘못된 로그인에 대해서는 403을 반환
+            return new ResponseEntity<>("아이디 또는 비밀번호가 잘못되었습니다.", HttpStatus.FORBIDDEN);
+        }
+        //서버에 저장된 id-pw 쌍과 일치한다면 토큰을 발급
+        Member member = memberService.getMemberByEmail(memberRequestDto.email()).get();
+        String token = "Bearer " + jwtAuthService.createJwt(member.getEmail(), member.getMemberId());
+        response.addHeader("Authorization", token);
         return ResponseEntity.ok().body(new JwtResponseDto(token));
     }
