--- src/main/java/gift/interceptor/RequestWrapper.java
@@ -1,17 +1,0 @@
-package gift.interceptor;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletRequestWrapper;
-
-public class RequestWrapper extends HttpServletRequestWrapper {
-
-    /**
-     * Constructs a request object wrapping the given request.
-     *
-     * @param request The request to wrap
-     * @throws IllegalArgumentException if the request is null
-     */
-    public RequestWrapper(HttpServletRequest request) {
-        super(request);
-    }
-}
